requirements.txt
Package                Version   
---------------------- --------- 
absl-py                0.11.0
astunparse             1.6.3
backcall               0.2.0
cachetools             4.1.1
certifi                2020.12.5
chardet                3.0.4
cloudpickle            1.6.0
cycler                 0.10.0
Cython                 0.29.21
decorator              4.4.2
detectron2             0.1      
flatbuffers            1.12
future                 0.18.2
fvcore                 0.1.2
gast                   0.3.3
gensim                 3.8.3
google-auth            1.23.0
google-auth-oauthlib   0.4.2
google-pasta           0.2.0
grpcio                 1.32.0
h5py                   2.10.0
idna                   2.10
imagesize              1.2.0
importlib-metadata     3.1.1
ipython                7.16.1
ipython-genutils       0.2.0
jedi                   0.17.2
joblib                 1.0.0
Keras-Preprocessing    1.1.2
kiwisolver             1.3.1
Markdown               3.3.3
matplotlib             3.3.3
networkx               2.5
node2vec               0.4.0
numpy                  1.19.4
oauthlib               3.1.0
opencv-python          4.4.0.46
opt-einsum             3.3.0
pandas                 1.1.5
parso                  0.7.1
pexpect                4.8.0
pickleshare            0.7.5
Pillow                 8.0.1
pip                    20.3.1
plotly                 4.14.3
portalocker            2.0.0
prompt-toolkit         3.0.8
protobuf               3.14.0
ptyprocess             0.6.0
pyasn1                 0.4.8
pyasn1-modules         0.2.8
pycocotools            2.0
pydot                  1.4.1
Pygments               2.7.2
pyparsing              2.4.7
python-dateutil        2.8.1
pytz                   2020.5
PyYAML                 5.3.1
requests               2.25.0
requests-oauthlib      1.3.0
retrying               1.3.3
rsa                    4.6
scikit-learn           0.24.0
scipy                  1.5.4
setuptools             50.3.2
six                    1.15.0
smart-open             4.0.1
stellargraph           1.2.1
tabulate               0.8.7
tensorboard            2.4.0
tensorboard-plugin-wit 1.7.0
tensorflow             2.4.0
tensorflow-estimator   2.4.0
termcolor              1.1.0
threadpoolctl          2.1.0
torch                  1.4.0
torchvision            0.5.0
tqdm                   4.54.1
traitlets              4.3.3
typing-extensions      3.7.4.3
urllib3                1.26.2
wcwidth                0.2.5
Werkzeug               1.0.1
wheel                  0.36.1
wrapt                  1.12.1
yacs                   0.1.8
zipp                   3.4.0

